requirements.txt
torch_efficient_distloss==0.1.3
torch_scatter==2.0.9
pytorch-lightning==1.8.0
torchmetrics==0.9.3
gin-config==0.5.0
numpy==1.24.3
scikit-image==0.19.3
imageio==2.35.1
imageio-ffmpeg==0.4.7
opencv-python==4.6.0.66
pillow==9.0.1
matplotlib==3.2.1
tifffile==2023.7.10
gdown==4.5.1
piq==0.8.0
piqa==1.2.2
openexr==3.2.4
tensorboard==2.14.0
tqdm==4.62.3
ninja==1.10.2.3
requests==2.27.1
configargparse==1.5.3
wandb==0.13.2
functorch==0.1.1
tensorboard==2.14.0